+++ dice_maiden.rb
@@ -1,6 +1,6 @@
 # Dice bot for Discord
 # Author: Humblemonk
-# Version: 7.2.0
+# Version: 8.4.0
 # Copyright (c) 2017. All rights reserved.
 # !/usr/bin/ruby
 # If you wish to run a single instance of this bot, please follow the "Manual Install" section of the readme!
@@ -16,9 +16,8 @@ require 'sqlite3'
 Dotenv.load
 @total_shards = ENV['SHARD'].to_i
 # Add API token
-@bot = Discordrb::Bot.new token: ENV['TOKEN'], num_shards: @total_shards, shard_id: ARGV[0].to_i,
-                          intents: %i[servers server_messages direct_messages], ignore_bots: true, fancy_log: true
-@bot.gateway.check_heartbeat_acks = false
+@bot = Discordrb::Commands::CommandBot.new token: ENV['TOKEN'], num_shards: @total_shards, shard_id: ARGV[0].to_i,
+                                           intents: %i[servers], ignore_bots: true, fancy_log: true
 @shard = ARGV[0].to_i
 @launch_option = ARGV[1].to_s
 @prefix = ''
@@ -35,22 +34,36 @@ end
 
 mutex = Mutex.new
 
-@bot.message do |event|
+if @shard == 0
+  puts "Shard #{@shard} is registering commands"
+  @bot.register_application_command(:roll, 'Ask Dice Maiden to roll some dice!') do |cmd|
+    cmd.string('message', 'roll syntax sent to Dice Maiden. Type help or visit github to view possible commands', required: true)
+  end
+
+  @bot.register_application_command(:r, 'Ask Dice Maiden to roll some dice!') do |cmd|
+    cmd.string('message', 'roll syntax sent to Dice Maiden. Type help or visit github to view possible commands', required: true)
+  end
+end
+
+inc_cmd = lambda do |event|
   # Locking the thread to prevent messages going to the wrong server
   mutex.lock
   begin
-    # handle !dm <command>
-    next if check_server_options(event) == true
+    @event_roll = event.options.values.join('')
+    # handle !dm <command>. DEPRECATED WITH SLASH COMMANDS
+    # next if check_server_options(event) == true
 
-    # check the server request options
-    check_request_option(event)
+    # check what prefix the server should be using. DEPRECATED WITH SLASH COMMANDS
+    # check_prefix(event)
+    # check if input is even valid. DEPRECATED WITH SLASH COMMANDS
+    # next if input_valid(event) == false
+    # check_request_option(event)
 
-    # check what prefix the server should be using
-    check_prefix(event)
-    # check if input is even valid
-    next if input_valid(event) == false
+    # check for comment
+    check_comment
+    @roll_request = @event_roll.dup
 
-    @input = alias_input_pass(event.content) # Do alias pass as soon as we get the message
+    @input = alias_input_pass(@event_roll) # Do alias pass as soon as we get the message
     @simple_output = false
     @wng = false
     @dh = false
@@ -64,20 +77,19 @@ mutex = Mutex.new
     @roll_set = nil
     next unless roll_sets_valid(event)
 
-    if @input =~ /^\s*d/
-      roll_to_one = @input.lstrip
-      roll_to_one.prepend('1')
-      @input = roll_to_one
-    end
+    # check for single dice rolls
+    @input.gsub!(%r{(?<!\d)(^|[+*/-]\s?)d(\d+)}, '\11d\2') if @input.match?(%r{(?<!\d)(^|[+*/-]\s?)d(\d+)})
 
     @roll = @input
     @check = @prefix + @roll
-    @comment = ''
     @test_status = ''
     # check user
     check_user_or_nick(event)
-    # check for comment
-    check_comment
+    # check for empty roll
+    if @event_roll.empty?
+      event.respond(content: "#{@user} roll is empty! Please type a complete dice roll message")
+      next
+    end
     # check for modifiers that should apply to everything
     check_universal_modifiers
 
@@ -117,9 +129,9 @@ mutex = Mutex.new
 
         log_roll(event) if @launch_option == 'debug'
         if @comment.to_s.empty? || @comment.to_s.nil?
-          event.respond "#{@user} Rolls:\n#{@roll_set_results}"
+          event.respond(content: "#{@user} Request: `[#{@roll_request.strip}]` Rolls:\n#{@roll_set_results}")
         else
-          event.respond "#{@user} Rolls:\n#{@roll_set_results} Reason: `#{@comment}`"
+          event.respond(content: "#{@user} Rolls:\n#{@roll_set_results} Reason: `#{@comment}`")
         end
         next
       end
@@ -135,7 +147,7 @@ mutex = Mutex.new
       if check_wrath == true
         respond_wrath(event, dnum)
       else
-        event.respond build_response
+        event.respond(content: build_response)
         check_fury(event)
       end
     end
@@ -148,19 +160,23 @@ mutex = Mutex.new
     # Simplify roll and send it again if we error out due to character limit
     if (e.message.include? 'Message over the character limit') || (e.message.include? 'Invalid Form Body')
       if @roll_set.nil?
-        event.respond "#{@user} Roll #{@dice_result} Reason: `Simplified roll due to character limit`"
+        event.respond(content: "#{@user} Roll #{@dice_result} Reason: `Simplified roll due to character limit`")
       else
-        event.respond "#{@user} Rolls:\n#{@error_check_roll_set}Reason: `Simplified roll due to character limit`"
+        event.respond(content: "#{@user} Rolls:\n#{@error_check_roll_set}Reason: `Simplified roll due to character limit`")
       end
-    elsif (e.message.include? "undefined method `join' for nil:NilClass") || (e.message.include? "The bot doesn't have the required permission to do this!") || (e.message.include? '500: Internal Server Error')
+    elsif (e.message.include? "undefined method `join' for nil:NilClass") || (e.message.include? "The bot doesn't have the required permission to do this!") || (e.message.include? '500: Internal Server Error') || (e.message.include? '500 Internal Server Error')
       time = Time.now.getutc
       File.open('dice_rolls.log', 'a') { |f| f.puts "#{time} ERROR: #{e.message}" }
     else
-      event.respond('Unexpected exception thrown! (' + e.message + ")\n\nPlease drop us a message in the #support channel on the dice maiden server, or create an issue on Github.")
+      event.respond(content: ('Unexpected exception thrown! (' + e.message + ")\n\nPlease drop us a message in the #support channel on the dice maiden server, or create an issue on Github."))
     end
   end
   mutex.unlock
 end
+
+@bot.application_command(:roll, &inc_cmd)
+@bot.application_command(:r, &inc_cmd)
+
 if @launch_option == 'lite'
   @bot.run
 else
@@ -168,7 +184,7 @@ else
 
   # Sleep until bot is ready and then set listening status
   sleep(1) until @bot.ready
-  @bot.update_status('online', '!roll', nil, since = 0, afk = false, activity_type = 2)
+  @bot.update_status('online', '/roll', nil, since = 0, afk = false, activity_type = 2)
   # Check every 5 minutes and log server count
   loop do
     sleep 300