+++ earthdawn_logic.rb
@@ -24,12 +24,12 @@
 ]
 
 def replace_earthdawn(event)
-  roll = @input.match(/#{@prefix}\sed(\d+)/i)
+  roll = @input.match(/^\s*ed(\d+)/i)
   step = roll[1].to_i
   if step.between?(1, 20)
     @input.sub!('ed' + roll[1], @earthdawn_replacements[step])
   else
-    event.respond 'Only steps 1-20 are implemented'
+    event.respond(content: 'Only steps 1-20 are implemented')
     false
   end
 end